Move transformer enums into common
